+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 [PR #127](https://github.com/apollostack/apollo-server/pull/127)
 * Camelcase Hapi.  Issue #129. ([@nnance](https://github.com/nnance)) in
 [PR #132](https://github.com/apollostack/apollo-server/pull/132)
+* Fix error handling when parsing variables parameter. Issue #130. ([@nnance](https://github.com/nnance)) in
+[PR #131](https://github.com/apollostack/apollo-server/pull/131)
 
 ### v0.2.6
 * Expose the OperationStore as part of the public API. ([@nnance](https://github.com/nnance))